+++ src/components/Produtos.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { ListaProdutos } from '../logic/produtos'
+
 import styled from 'styled-components'
 
 
@@ -8,6 +8,56 @@ const ImagemProduto = styled.img`
 `
 
 export default class Produtos extends React.Component {
+
+    state = {
+
+        ListaProdutos = [
+            {
+                imagemProduto: "https://blog.enem.com.br/wp-content/uploads/2020/04/img-materia-r7-inter.jpg",
+                nomeProduto: "Planeta Miller",
+                valor: 85,
+                id: 1
+            },
+            
+            {
+                imagemProduto: "https://gizmodo.uol.com.br/wp-content/blogs.dir/8/files/2019/04/buraco-negro-gargantua-970x444.jpg",
+                nomeProduto: "Buraco Negro Gargantua",
+                valor: 90,
+                id: 2
+            },
+            
+            {
+                imagemProduto: "https://blogdoishii.files.wordpress.com/2018/07/k01vra.jpg",
+                nomeProduto: "Planeta Ranger",
+                valor: 55,
+                id: 3
+            },
+            {
+                imagemProduto: "https://imagens.publico.pt/imagens.aspx/885789?tp=UH&db=IMAGENS&type=JPG",
+                nomeProduto: "Planeta Pandora",
+                valor: 75,
+                id: 4
+            },
+            
+            {
+                imagemProduto: "https://super.abril.com.br/wp-content/uploads/2020/01/agua-marte-curiosity.jpg",
+                nomeProduto: "Planeta Marte",
+                valor: 20,
+                id: 5
+            },
+            
+            {
+                imagemProduto: "https://cdn.universoracionalista.org/wp-content/uploads/2020/06/ceres-dawn.jpg",
+                nomeProduto: "Asteróide B-612",
+                valor: 32,
+                id: 6
+            }
+        ]
+
+    }
+
+   
+    
 
     renderizaProdutos = () => {
         let lista = [];
@@ -23,6 +73,13 @@ export default class Produtos extends React.Component {
         }
 
         return lista;
+    }
+
+    somarItem = (itemId) => {
+        this.ValorTotal += this.ListaProdutos[itemId-1].valor;
+        this.ProdutosNoCarrinho.push(ListaProdutos[itemId-1]);
+        
+        return this.ValorTotal;
     }
 
     render() {
